[PCFG-184] fix
