+++ src/main/resources/h2/data-h2.sql
@@ -67,7 +67,7 @@ INSERT INTO NODO4_CFG.STAZIONI (OBJ_ID, ID_STAZIONE, ENABLED, IP, NEW_PASSWORD, 
                                 PORTA_4MOD, SERVIZIO_4MOD, PROXY_ENABLED, PROXY_HOST, PROXY_PORT, PROXY_USERNAME,
                                 PROXY_PASSWORD, PROTOCOLLO_AVV, IP_AVV, PORTA_AVV, SERVIZIO_AVV, TIMEOUT, NUM_THREAD,
                                 TIMEOUT_A, TIMEOUT_B, TIMEOUT_C, FLAG_ONLINE, VERSIONE, SERVIZIO_NMP)
-VALUES (3, '1112', 'Y', 'site.eu', null, 'password', 80, 'HTTP',
+VALUES (3, '1123', 'Y', 'site.eu', null, 'password', 80, 'HTTP',
         'site.it', 'page.html', 443, null,
         'service/', 'Y', 'service/', 2, 'HTTPS', 'HTTP', null, null, null,
         'Y', '10.101.1.95', 8080, null, null, 'HTTP', null, null, null, 120, 2, 15, 30, 120, 'Y', 1, null);
