+++ src/test/resources/file/ci_station_ok.csv
@@ -1,3 +1,3 @@
 cf,idstazione,ambiente,broadcast,auxdigit,codicesegregazione,applicationcode,datavalidita,operazione
-paStress65,stazPaStress65,COLL,N,3,47,,2022-03-21,A
-12345678900,stazioneLorenz,ESER,N,3,47,,2022-03-21,A
+1110001,80007580279_01,COLL,N,3,47,,2022-03-21,A
+00168480242,1123,ESER,N,3,47,,2022-03-21,A
